requirements.txt
streamlit==1.5.1
pandas==1.3.5
base58==2.1.1
numpy==1.19.5
pillow==8.4.0
scikit-learn==1.0.1
